wallet: remove file `psbt.go`
We now move all helper methods into `psbt_manager.go`.
